rollback to commit

+++ workspace-gateway/src/main/java/com/aliware/tianchi/remote/Access.java
@@ -1,7 +1,5 @@
 package com.aliware.tianchi.remote;
 
-import com.aliware.tianchi.util.ScalableSemaphore;
-
 import java.util.List;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
@@ -10,7 +8,7 @@ import java.util.stream.Collectors;
 public class Access {
     public static Map<String, Status> providerMap;
 
-    public static Map<String, Integer> maxAvailableThreads = new ConcurrentHashMap<>();
+    public static Map<String, Double> maxAvailableThreads = new ConcurrentHashMap<>();
 
     public static List<Double> getDuration() {
         return providerMap.values().stream().map(Status::getCurDuration).collect(Collectors.toList());
